+++ headerFile/Monster.h
@@ -3,7 +3,6 @@
 #include "Item.h"
 #include "Character.h"
 #include "MonsterData.h"
-#include "MyMath.h"
 #include "DecoWord.h"
 
 using namespace std;
